Replace AutoregistrationSpec with AutoregistrationTests. For now, we just commented out 2 unit tests testing precondition calls since those cannot be tested without depending on Nimble or CwlPreconditionTesting.
